requirements.txt
# Dependências principais do curso LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
python-dotenv>=1.0.0

# ===== OPÇÕES GRATUITAS (ESCOLHA UMA) =====

# Opção 1: Ollama (modelos locais - 100% gratuito)
# Instale Ollama em: https://ollama.ai
# Depois execute: ollama pull llama2

# Opção 2: Hugging Face (modelos na nuvem - gratuito com limites)
huggingface_hub>=0.19.0

# Opção 3: OpenAI (pago - melhor qualidade)
# langchain-openai>=0.0.5
# openai>=1.0.0

# ===== DEPENDÊNCIAS OPCIONAIS =====

# Para document loaders
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
youtube-transcript-api>=0.6.0

# Para vector stores
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Para agents e ferramentas
wikipedia>=1.4.0
duckduckgo-search>=4.1.0

# Para deploy e interfaces
streamlit>=1.28.0
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Para processamento de dados
pandas>=2.0.0
numpy>=1.24.0

# Para visualização
matplotlib>=3.7.0
seaborn>=0.12.0

# Para testes e desenvolvimento
jupyter>=1.0.0
ipykernel>=6.25.0 